Refactored the code

+++ src/features/cart/cartSlice.js
@@ -1,6 +1,17 @@
 import { createSlice } from '@reduxjs/toolkit';
+
 const initialState = {
   cart: [],
+
+  // cart: [
+  //   {
+  //     pizzaId: 12,
+  //     name: 'Mediterranean',
+  //     quantity: 2,
+  //     unitPrice: 16,
+  //     totalPrice: 32,
+  //   },
+  // ],
 };
 
 const cartSlice = createSlice({
@@ -8,36 +19,39 @@ const cartSlice = createSlice({
   initialState,
   reducers: {
     addItem(state, action) {
-      //payload = newItem
+      // payload = newItem
       state.cart.push(action.payload);
     },
     deleteItem(state, action) {
-      //payload = pizzaId
+      // payload = pizzaId
       state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
     },
-    increaseItemQunantity(state, action) {
-      //payload = pizzaId
+    increaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
+
       item.quantity++;
       item.totalPrice = item.quantity * item.unitPrice;
     },
     decreaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
+
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
 
-      if (!item.quantity === 0)
-        cartSlice.caseReducers.deleteItem(state, action);
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
     clearCart(state) {
       state.cart = [];
     },
   },
 });
+
 export const {
   addItem,
   deleteItem,
-  increaseItemQunantity,
+  increaseItemQuantity,
   decreaseItemQuantity,
   clearCart,
 } = cartSlice.actions;
@@ -54,3 +68,5 @@ export const getTotalCartPrice = (state) =>
 
 export const getCurrentQuantityById = (id) => (state) =>
   state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
+
+// 'reselect'
